Add Early Distribution Scaling Functionality
This enables scaling the input distribution to the desired range and not the generated Values.
Therefore the generated Values are more realistical after sacling as the minimum and maximum are no longer equal to the input scale.
This is acieved via the already for the scaling used Method with the differance that it is applied early to the orignal distribution.