Fix random seed for interaction generation
This adds a flag in the model if a random seed is set as well as a random seed attribute.
The seed is used by the AbstractWeaver.cs and the FMScaling.cs files.
The seed now works for feature/interaction values and the found interactions.
This does not completly remove all randomness, especially for the Generic Algorithm.