Fix bug with the Interaction Generation
The method checking if an interaction allready exist was flawed.
The check was if the new interaction containes all features of an existing interaction which fails if the existing interaction is a subset of the features of the new interaction.
The new interaction can not be a subset of an existing interaction as interactions are generated starting with the smaller order.