Add relative interaction count option to the interaction_count command
The new command replaces the absolute interaction number parameter with a relative percentage number for each interaction order:
set interaction_count relative_0.5#0.3#1
Each order is separated via #. So for this example 50% of all interactions of order 1 and 30% of all interactions of order 2 and all interactions of order 3 are generated.
As Thor needs the concrete interaction number upfront, all interactions of the required orders are generated and used to calculate the concrete numbers.
Furthermore the selected interactions are stored to avoid generating them twice.